Move map commands out of folder
