+++ src/commands/arenas/index.js
