+++ src/commands/battleRoyale/index.js
